requirements/tests.txt
argparse
coverage
django-discover-runner
mock == 1.0.1
tox
